+++ README.md
@@ -6,7 +6,7 @@ It consists of
 
 1. Transport.js abstract class that is used by K3poControl.js to connect to the [K3po Driver](https://github.com/k3po/k3po/tree/develop/driver)
 1. BBoshTransport.js, a streaming [Binary Bidirectional Streams Over HTTP transport protocol](https://github.com/k3po/k3po/tree/develop/specification/bbosh) that implements K3poControlTransport and that can be used in the browser on which to connect to the [K3po Driver](https://github.com/k3po/k3po/tree/develop/driver).
-1. TcpTransport.js, a node tcp implementation of Transport.js that can be used to control k3po from node
+1. TcpTransportFactory.js, a node tcp implementation of Transport.js that can be used to control k3po from node
 1. K3poControl.js, a [k3po control protocol](https://github.com/k3po/k3po/tree/develop/specification/k3po.control) that is used by talk to the [K3po Driver](https://github.com/k3po/k3po/tree/develop/driver) via a K3poTransport that provides an API for starting/stopping k3po scripts
 1. mocha-k3po.js, a [mocha](https://mochajs.org/) plugin that can control k3po scripts as part of the test framework
