--- test/TestTransport.js
@@ -1,16 +1,0 @@
-var ControlTransportApi = require('../src/base/ControlTransportApi.js').ControlTransportApi;
-
-/**
- * Test transport used as mock for testing
- * @constructor
- */
-function TestTransport(url) {
-    ControlTransportApi.call(this);
-    this.url = url;
-}
-
-TestTransport.prototype = Object.create(ControlTransportApi.prototype);
-
-TestTransport.prototype.constructor = ControlTransportApi;
-
-module.export = TestTransport;
